requirements.txt
fastapi==0.103.2
numpy
pillow
python-multipart
scikit-learn==0.22.2
torch
torchcam
torchvision
# utils
uvicorn